Exception handling pending

+++ src/main/java/com/example/spring/service/CategoryServiceImpl.java
@@ -8,36 +8,48 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class CategoryServiceImpl implements CategoryService {
 
-    @Autowired
     private CategoryRepository categoryRepository;
+
+    @Autowired
+    CategoryServiceImpl(CategoryRepository categoryRepository) {
+        this.categoryRepository = categoryRepository;
+    }
+
 
     @Override
     public CategoryDTO saveCategory(CategoryDTO categoryDTO) {
+
         Category category = CategoryMapper.getCategory(categoryDTO);
         category = categoryRepository.save(category);
         return CategoryMapper.getCategoryDTO(category);
+
     }
 
     @Override
     public List<CategoryDTO> getAllCategory() {
-        List<Category> categories = categoryRepository.findAll();
-        return categories.stream().map(category -> CategoryMapper.getCategoryDTO(category)).toList();
+        List<Category> categoryList = categoryRepository.findAll();
+        return categoryList.stream().map(sp -> CategoryMapper.getCategoryDTO(sp)).toList();
     }
 
     @Override
     public CategoryDTO getCategoryById(Long id) {
-        Category category = categoryRepository.findById(id).orElseThrow(() -> new RuntimeException("Caterogy not found"));
+        Category category = categoryRepository.findById(id).orElseThrow(() -> new RuntimeException("Category not found"));
         return CategoryMapper.getCategoryDTO(category);
-
     }
 
     @Override
-    public void deleteCategory(Long id) {
-        categoryRepository.deleteById(id);
+    public boolean deleteCategoryById(Long id) {
+        Optional<Category> isAvailable = categoryRepository.findById(id);
+        if (isAvailable.isPresent()) {
+            categoryRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
